Write general internal condensing function
